Remove Youtube broad scope
The scope removed was asking the user permissions to manage the FULL Youtube account and that was not needed. In fact the read.only scope is more than enough to get the permission token to view the user plugin. The extra scope has been removed because when submitting a custom google app for verification, Google will complain of this large scope (and this also gives too much power to this plugin, with the risk of compromising the youtube account of the user)